nuevos eventos y fetch

+++ JavaScript/script.js
@@ -11,16 +11,9 @@ class Producto {
 }
 
 function agregarProducto(nombre, precio, img) {
-    let nuevoProducto = new Producto(nombre, precio, img)
+    let nuevoProducto = new Producto(nombre, precio, img);
     productos.push(nuevoProducto)
 }
-
-agregarProducto("Remera", 15000, "https://cdn.pixabay.com/photo/2023/05/06/01/34/t-shirt-7973404_1280.jpg")
-agregarProducto("Gorra", 5000, "https://cdn.pixabay.com/photo/2017/07/21/14/28/hat-2525910_1280.png")
-agregarProducto("Jean", 20000, "https://cdn.pixabay.com/photo/2018/10/10/14/23/pants-3737399_1280.jpg")
-agregarProducto("Medias", 3000, "https://cdn.pixabay.com/photo/2015/04/09/19/56/sock-715024_960_720.jpg")
-agregarProducto("Campera", 40000, "https://cdn.pixabay.com/photo/2017/09/07/04/54/khaki-2723896_1280.jpg")
-agregarProducto("Zapatillas", 35000, "https://cdn.pixabay.com/photo/2020/07/15/18/24/footwear-5408640_960_720.png")
 
 const contProductos = document.querySelector(".cont-productos")
 
@@ -38,17 +31,44 @@ const mostrarProductos = (listaProductos) => {
         `;
         contProductos.appendChild(tarjetaProducto);
     })
+
+    const botonProd = document.querySelectorAll(".boton-prod")
+
+    botonProd.forEach(btn =>
+        btn.addEventListener("click", (event) =>
+            agregarCarrito(event.target.id),
+        )
+    )
+
+    botonProd.forEach(btn =>
+        btn.addEventListener("click", () => {
+            Swal.fire({
+                toast: true,
+                position: "top-end",
+                icon: "success",
+                showConfirmButton: false,
+                timer: 1200,
+                title: "¡Agregado al carrito con éxito!"
+            })
+        }))
 }
 
-mostrarProductos(productos)
-
-const botonProd = document.querySelectorAll(".boton-prod")
-
-botonProd.forEach(btn =>
-    btn.addEventListener("click", (event) =>
-        agregarCarrito(event.target.id)
-    )
-)
+const productosPreexistentes = async () => {
+    if (productos.length === 0) {
+        try {
+            const URLProductos = "./productos.JSON";
+            const productosBase = await fetch(URLProductos);
+            const productosBaseJSON = await productosBase.json()
+            productosBaseJSON.forEach(producto => {
+                agregarProducto(producto.nombre, producto.precio, producto.img)
+            })
+        } catch (err) {
+            console.error("Se produjo un error al obtener los productos:", err)
+        } finally {
+            mostrarProductos(productos)
+        }
+    }
+}
 
 let carrito = []
 
@@ -59,51 +79,71 @@ const convertirCarrito = (lista) => {
 
 const encontrarCarrito = () => {
     const carritoLocal = localStorage.getItem("carritoJSON");
-    if (carritoLocal === "") { } else {
+    if (carritoLocal === "" || carritoLocal === null) { } else {
         carrito = JSON.parse(carritoLocal);
     }
 }
 
+const mostrarCarrito = document.querySelector(".texto-carrito")
+const finCompra = document.querySelector(".fin-compra");
+const vaciarCarrito = document.querySelector(".vaciar-carrito")
+
+function totalCarrito() {
+    let total = 0;
+    if (carrito !== null || carrito !== "") {
+        for (i = 0; i < carrito.length; i++) {
+            total += carrito[i].precio * carrito[i].cantidad
+        }
+    }
+    mostrarCarrito.innerText = `El total de tu carrito es $${total}`;
+}
+
 function agregarCarrito(id) {
     let productoCarrito = productos.find(prod => prod.id === parseInt(id));
-    if (carrito.some(prod => prod.id == parseInt(id))) {
+    if (carrito.some(prod => prod.id === parseInt(id))) {
         productoCarrito.cantidad++
     } else {
         productoCarrito.cantidad = 1;
         carrito.push(productoCarrito);
     }
+
     convertirCarrito(carrito)
     totalCarrito()
 }
 
-const mostrarCarrito = document.querySelector(".texto-carrito")
-
-function totalCarrito() {
-    let total = 0;
-    for (i = 0; i < carrito.length; i++) {
-        total += carrito[i].precio * carrito[i].cantidad
-    }
-    mostrarCarrito.innerText = `El total de tu carrito es $${total}`;
-}
-
-const finCompra = document.querySelector(".fin-compra");
-const vaciarCarrito = document.querySelector(".vaciar-carrito")
-
 finCompra.addEventListener("click", () => {
-    alert("Muchas gracias por tu compra!")
+    Swal.fire({
+        title: "¡Muchas gracias por tu compra!",
+        icon: "success",
+        confirmButtonColor: "#6e85a3"
+    });
     carrito = [];
     localStorage.setItem("carritoJSON", "")
     totalCarrito();
 })
 
 vaciarCarrito.addEventListener("click", () => {
-    carrito = [];
-    localStorage.setItem("carritoJSON", "")
-    totalCarrito();
+    Swal.fire({
+        title: "¿Desea vaciar el carrito de compras?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#6e85a3",
+        confirmButtonText: "Aceptar",
+        cancelButtonText: "Cancelar"
+    }).then((result) => {
+        if (result.isConfirmed) {
+            carrito = [];
+            localStorage.setItem("carritoJSON", "");
+            totalCarrito()
+        }
+    });
 })
 
 const app = () => {
+    productosPreexistentes()
     encontrarCarrito()
+    convertirCarrito(carrito)
     totalCarrito()
 }
